# Introduce thread pool for writing goto binaries in parallel