--- a/cprover_bindings/src/lib.rs
+++ b/cprover_bindings/src/lib.rs
@@ -31,6 +31,7 @@
 
 #![feature(f128)]
 #![feature(f16)]
+#![feature(negative_impls)]
 
 mod env;
 pub use env::global_dead_object;
@@ -41,4 +42,5 @@ pub mod utils;
 pub use irep::serialize;
 pub use machine_model::{MachineModel, RoundingMode};
 mod cbmc_string;
+pub use cbmc_string::sync::{InternerSpecific, WithInterner};
 pub use cbmc_string::{InternString, InternStringOption, InternedString};