@tautschnig 2025-07-25
Running `kani autoharness -Z autoharness --list ...` for the standard library this command previously took 2:05 hours of wall-clock time (and 11300 seconds of CPU time) on an r5.metal. With the changes in this PR and increasing `NUM_FILE_EXPORT_THREADS` to 32 (can we please use the value of `--jobs` for that?) the same command took 1:25 hours (approximately 25% improvement) despite using 13000 seconds of CPU time (some overhead is not unexpected).