@tautschnig 2025-07-28
> @tautschnig Does it make sense to cap the number of threads used (taken from the value of `--jobs`) at `# of proof harnesses - 1`? For workloads of just a single proof this will force the main thread to just handle everything itself and potentially negate this regression.

Yes, this capping sounds like a good idea!