@AlexanderPortland 2025-08-01
Just ran the tests and, on my local machine, this change brings the `prost` crate's end to end compile time down 45%. I've updated the PR description with the full #s from that benchmark and the standard library.